Backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-bcrypt
mysql-connector-python